requirements.txt
librosa
numpy
pyyaml
torch
